Atualização finalizando a parte do servidor

+++ Controller/TextMananger.cs
@@ -24,10 +24,10 @@ namespace XmlWebEditor.Controller
 {
     public class TextMananger
     {
-        public string VefiryXml(IFormFile Upload, IWebHostEnvironment environment, string fileName,string archiveName, ref string error)
+        public string VefiryXml(IFormFile Upload, IWebHostEnvironment environment, string file, ref string error)
         {
 
-            var file = Path.Combine(environment.ContentRootPath, archiveName, fileName + ".xml");
+            
             string xml = "";
             try
             {
@@ -42,8 +42,11 @@ namespace XmlWebEditor.Controller
                 XmlDocument document = new XmlDocument();
                 document.LoadXml(xml);
                 xml = XElement.Parse(xml).ToString();
-                var myFile = File.Create(file);
-                myFile.Close();
+                if (!File.Exists(file))
+                {
+                    var myFile = File.Create(file);
+                    myFile.Close();
+                }
                 using (var fileStream = new FileStream(file, FileMode.Create))
                 {
                     byte[] information = new UTF8Encoding(true).GetBytes(xml);
@@ -66,10 +69,8 @@ namespace XmlWebEditor.Controller
             }
 
         }//end XmlVerification
-        public string VerifyJson(IFormFile Upload, IWebHostEnvironment environment, string fileName,string archiveName, ref string error)
+        public string VerifyJson(IFormFile Upload, IWebHostEnvironment environment, string file, ref string error)
         {
-
-            var file = Path.Combine(environment.ContentRootPath, archiveName, fileName + ".json");
             string json = "";
             try
             {
@@ -103,15 +104,15 @@ namespace XmlWebEditor.Controller
                 return json;
             }
         }//end VerifyJson
-        public string SetFile(IFormFile Upload, IWebHostEnvironment environment, string fileName,string archiveName, ref string error)
+        public string SetFile(IFormFile Upload, IWebHostEnvironment environment, string file, ref string error)
         {
             try
             {
                 string contentType = Upload.ContentType;
                 if (contentType == "text/xml" || contentType == "application/xml")
-                    return VefiryXml(Upload, environment, fileName,archiveName, ref error);
+                    return VefiryXml(Upload, environment, file + ".xml", ref error);
                 else if (contentType == "text/json" || contentType == "application/json")
-                    return VerifyJson(Upload, environment, fileName, archiveName, ref error);
+                    return VerifyJson(Upload, environment,file + ".json", ref error);
                 else
                     throw new System.NullReferenceException();
             }
@@ -139,7 +140,7 @@ namespace XmlWebEditor.Controller
             aux = File.ReadAllText(newFile);
             return aux;
         }// end NewXmlFile
-        public string UpdateFile(IWebHostEnvironment environment, string fileName,string archiveName, string xmlText, ref string error)
+        public string UpdateFile(IWebHostEnvironment environment, string file, string xmlText, ref string error)
         {
             try
             {
@@ -147,11 +148,11 @@ namespace XmlWebEditor.Controller
                 if (xmlText.StartsWith("<"))
                 {
 
-                    return UpdateXmlFile(environment, fileName,archiveName, xmlText, ref error);
+                    return UpdateXmlFile(environment, file+".xml", xmlText, ref error);
                 }
                 else if (xmlText.StartsWith("{") || xmlText.StartsWith("["))
                 {
-                    return UpdateJsonFile(environment, fileName,  archiveName, xmlText, ref error);
+                    return UpdateJsonFile(environment, file+".json", xmlText, ref error);
                 }
                 else
                     throw new System.NullReferenceException();
@@ -163,7 +164,7 @@ namespace XmlWebEditor.Controller
             }
 
         }//end UpdateFile
-        public string UpdateXmlFile(IWebHostEnvironment environment, string fileName, string archiveName, string xmlText, ref string error)
+        public string UpdateXmlFile(IWebHostEnvironment environment, string file, string xmlText, ref string error)
         {
             string xml = xmlText;
             try
@@ -172,16 +173,11 @@ namespace XmlWebEditor.Controller
                 XmlDocument document = new XmlDocument();
                 document.LoadXml(xml);
                 xml = XElement.Parse(xml).ToString();
-                var file = Path.Combine(environment.ContentRootPath, archiveName, fileName + ".xml");
-                var fileAux = Path.Combine(environment.ContentRootPath, archiveName, "fileAux.xml");
-                File.WriteAllTextAsync(fileAux, xmlText);
+                File.WriteAllTextAsync(file, xmlText);
                 /*Resumidamente: File.WriteallTextAsync necessita de tempo
                  * para atualizar e então o document.load poder ser utilizado, por isto:
                  * */
                 //Thread.Sleep(100);
-                var myFile = File.Create(file);
-                myFile.Close();
-                File.Copy(fileAux, file, true);
                 return xml;
 
             }
@@ -198,7 +194,7 @@ namespace XmlWebEditor.Controller
                 return xml;
             }
         }//end UpdateXmlFile
-        public string UpdateJsonFile(IWebHostEnvironment environment, string fileName, string archiveName, string xmlText, ref string error)
+        public string UpdateJsonFile(IWebHostEnvironment environment, string file, string xmlText, ref string error)
         {
             string json = xmlText;
             dynamic jsonAux;
@@ -207,18 +203,13 @@ namespace XmlWebEditor.Controller
 
                 jsonAux = JsonConvert.DeserializeObject(json);
                 json = JsonConvert.SerializeObject(jsonAux, Newtonsoft.Json.Formatting.Indented);
-                var file = Path.Combine(environment.ContentRootPath, archiveName, fileName + ".json");
-                var fileAux = Path.Combine(environment.ContentRootPath, archiveName, "fileAux.json");
-                File.WriteAllTextAsync(fileAux, xmlText);
+                File.WriteAllTextAsync(file, xmlText);
                 /*Resumidamente: File.WriteallTextAsync necessita de tempo
                  * para atualizar e então o document.load poder ser utilizado, por isto:
                 */
                 //Thread.Sleep(100);
                 //obs: por enquanto aos testes não fora mais necessário, porém deixo aqui "em caso de" 
                
-                var myFile=File.Create(file);
-                myFile.Close();
-                File.Copy(fileAux, file, true);
                 return json;
 
             }
